+++ lib/goodmap.dart
@@ -7,11 +7,12 @@ export 'src/good_map_controller.dart'
         MarkerId,
         MarkerImage,
         MarkerOptions,
+        GlobePoint,
         PopupId,
         PopupOptions,
         PolylineId,
         PolylineOptions;
-export 'src/globe/globe_overlays.dart' show GlobePoint, GlobeArc;
+export 'src/globe/globe_overlays.dart' show GlobeArc;
 export 'src/globe/good_globe.dart' show GoodGlobe;
 export 'src/globe/good_map_globe.dart' show GoodMapGlobe;
 export 'src/good_map.dart' show GoodMap;